requirements.txt
asgiref==3.8.1
cloudinary==1.39.1
Django==4.2.11
django-cloudinary-storage==0.3.0
django-filter==24.2
djangorestframework==3.15.1
Pillow==8.2.0
pytz==2024.1
sqlparse==0.4.4
